refactor(import): route the GUI/CLI XG import through the ingest pipeline
Database.ImportXGMatch now delegates to ingest.MapXG + ingest.WriteMatch — the
same storage-backed pipeline the headless server uses — instead of its own
~750-line in-Database XG mapper. The duplicate-detection contract the GUI/CLI
rely on is preserved: an exact same-format re-import still returns
ErrDuplicateMatch (the ingest layer reports it as a silent skip, which the
delegating method translates back), while a cross-format canonical duplicate is
enriched in place and returns the existing match id without error.

This is the first step of unifying the two parallel import implementations
(database/db_import_* and ingest/*map). It required breaking a latent test
import cycle: the legacy↔ingest parity tests lived in package ingest and
imported package database, so once database imports ingest they would cycle.
Those parity tests had already done their job — they pass, proving the two
paths produce identical positions/analyses/hashes — so they are retired here.
The ingest XG path is now exercised end to end (with absolute-value assertions,
not just parity) by the database package's existing XG import tests
(count_positions, canonical_hash, xg_stats_reference, score_check,
mat_position), which all stay green through the delegated method.

Mechanics:
- Move the shared raw-readback test helpers (rawAnalyses, posData,
  normaliseAnalysisForCompare, positionKey) from xg_test.go to a new
  importsupport_test.go (no database import).
- Reseed db_test.go's native-.db importer tests via the ingest XGImporter
  instead of the legacy Database, dropping its database import.
- Delete the now-tautological/cycle-forming parity tests (xg, bgf, gnubg,
  position, position_bgf). BGF/GnuBG/position legacy paths are untouched and
  keep their database-package coverage until their own delegation PRs.
- Delete the 8 now-dead legacy XG mapping/save functions (db_import_xg.go
  1698 → 947 lines); drop the unused context/math imports.

Verified: go build ./..., go vet, golangci-lint (0 issues), and
go test -race ./... all green.
