+++ pkg/blunderdb/database/db_import_xg.go
@@ -1,22 +1,24 @@
 package database
 
 import (
-	"context"
 	"database/sql"
 	"fmt"
 	"log/slog"
-	"math"
 	"sort"
 	"strings"
 	"time"
 
+	"github.com/kevung/blunderdb/pkg/blunderdb/ingest"
 	"github.com/kevung/xgparser/xgparser"
 )
 
 // Match import and management functions
 
 // Import XG match file using xgparser library
-// This function uses raw segment parsing to capture complete cube action information
+// ImportXGMatch imports an eXtreme Gammon .xg match file. It delegates to the
+// shared ingest pipeline (ingest.MapXG → ingest.WriteMatch) — the same path the
+// headless server uses — persisting through the storage backend. The former
+// in-Database XG mapping it replaced now lives only inside the ingest package.
 func (d *Database) ImportXGMatch(filePath string) (int64, error) {
 	ctx, done := d.beginCancellableImport()
 	defer done()
@@ -24,206 +26,32 @@ func (d *Database) ImportXGMatch(filePath string) (int64, error) {
 	d.mu.Lock()
 	defer d.mu.Unlock()
 
-	// Parse the XG file using raw segments for complete data
-	imp := xgparser.NewImport(filePath)
-	segments, err := imp.GetFileSegments()
-	if err != nil {
-		return 0, fmt.Errorf("failed to get file segments: %w", err)
-	}
-
-	// Also parse lightweight structure for metadata
-	match, err := xgparser.ParseXG(segments)
-	if err != nil {
-		return 0, fmt.Errorf("failed to parse XG file: %w", err)
-	}
-
-	// Parse raw records for complete cube information
-	rawCubeInfo := make(map[string]*RawCubeAction) // key: "game_cubeIdx"
-	for _, seg := range segments {
-		if seg.Type == xgparser.SegmentXGGameFile {
-			records, _ := xgparser.ParseGameFile(seg.Data, -1)
-			gameNum := int32(0)
-			cubeIdx := 0
-			for _, rec := range records {
-				switch r := rec.(type) {
-				case *xgparser.HeaderGameEntry:
-					gameNum = r.GameNumber
-					cubeIdx = 0
-				case *xgparser.CubeEntry:
-					if r.Double != -2 { // Skip initial positions
-						key := fmt.Sprintf("%d_%d", gameNum, cubeIdx)
-						rawCubeInfo[key] = &RawCubeAction{
-							Double:   r.Double,
-							Take:     r.Take,
-							ActiveP:  r.ActiveP,
-							CubeB:    r.CubeB,
-							Position: r.Position,
-							Doubled:  r.Doubled,
-						}
-						cubeIdx++
-					}
-				}
-			}
-		}
-	}
-
-	// Parse match date
-	matchDate := parseMatchDate(match.Metadata.DateTime)
-
-	// Compute match hash for duplicate detection (includes full match transcription)
-	matchHash := ComputeMatchHash(match)
-
-	// Compute canonical hash (format-independent) for cross-format duplicate detection
-	canonicalHash := ComputeCanonicalMatchHashFromXG(match)
-
-	// Check for duplicate (same format) or canonical duplicate (cross-format)
-	canonicalMatchID, isCanonicalDuplicate, err := d.checkDuplicateMatchLocked(matchHash, canonicalHash)
+	graph, err := ingest.MapXG(filePath)
 	if err != nil {
 		return 0, err
 	}
-
-	// Begin transaction for atomic import
-	tx, err := d.db.Begin()
+	tx, err := d.store.BeginTx(ctx)
 	if err != nil {
-		return 0, fmt.Errorf("failed to begin transaction: %w", err)
+		return 0, err
 	}
-	// Unconditional rollback: a no-op once tx.Commit() succeeds, but guarantees
-	// the transaction is released on any early return — including ctx
-	// cancellation, whose error propagates through a shadowed `err` and so
-	// would not trip a conditional `if err != nil` rollback.
-	defer tx.Rollback()
-
-	// Insert match metadata (including match_hash and canonical_hash)
-	// If canonical duplicate, don't create new match - reuse existing match ID
-	var matchID int64
-	if isCanonicalDuplicate {
-		matchID = canonicalMatchID
-		slog.Info("canonical duplicate detected, reusing match", "matchID", matchID)
-	} else {
-		result, err := tx.Exec(`
-			INSERT INTO match (player1_name, player2_name, event, location, round, 
-			                   match_length, match_date, file_path, game_count, match_hash, canonical_hash)
-			VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-		`, match.Metadata.Player1Name, match.Metadata.Player2Name,
-			match.Metadata.Event, match.Metadata.Location, match.Metadata.Round,
-			match.Metadata.MatchLength, matchDate, filePath, len(match.Games), matchHash, canonicalHash)
-
-		if err != nil {
-			return 0, fmt.Errorf("failed to insert match: %w", err)
-		}
-
-		matchID, err = result.LastInsertId()
-		if err != nil {
-			return 0, fmt.Errorf("failed to get match ID: %w", err)
-		}
-
-		autoLinkTournament(tx, matchID, match.Metadata.Event)
+	res, err := ingest.WriteMatch(ctx, tx, "", graph, nil)
+	if err != nil {
+		_ = tx.Rollback()
+		return 0, err
 	}
-
-	// Build a per-import deduplication cache keyed by Zobrist hash.
-	// The preload of all existing positions is no longer needed — the UNIQUE index
-	// on position.zobrist_hash handles cross-import dedup via INSERT OR IGNORE.
-	cache := newImportCache()
-
-	if isCanonicalDuplicate {
-		// Canonical duplicate: import analysis to existing positions, create genuinely new ones
-		for _, game := range match.Games {
-			var pendingCubeComment string // Comment from a skipped "No Double" cube decision
-			for i, move := range game.Moves {
-				// Detect "No Double" cube decisions that will be skipped
-				if move.MoveType == "cube" && move.Comment != "" {
-					isSkipped := move.CubeMove == nil || move.CubeMove.Analysis == nil
-					if isSkipped {
-						pendingCubeComment = move.Comment
-					}
-				}
-
-				if move.MoveType == "checker" && move.CheckerMove != nil {
-					// Carry forward comment from a skipped "No Double" cube decision
-					if pendingCubeComment != "" {
-						if move.Comment == "" {
-							game.Moves[i].Comment = pendingCubeComment
-							move.Comment = pendingCubeComment
-						} else {
-							combined := pendingCubeComment + "\n" + move.Comment
-							game.Moves[i].Comment = combined
-							move.Comment = combined
-						}
-						pendingCubeComment = ""
-					}
-
-					pos, err := d.createPositionFromXG(move.CheckerMove.Position, &game, int32(match.Metadata.MatchLength), 0, move.CheckerMove.ActivePlayer)
-					if err != nil {
-						continue
-					}
-					pos.PlayerOnRoll = convertXGPlayerToBlunderDB(move.CheckerMove.ActivePlayer)
-					pos.DecisionType = CheckerAction
-					pos.Dice = [2]int{int(move.CheckerMove.Dice[0]), int(move.CheckerMove.Dice[1])}
-
-					posID, err := d.findOrCreatePositionForCanonicalDuplicate(tx, pos, cache)
-					if err != nil {
-						continue
-					}
-
-					// Save checker analysis to position
-					if len(move.CheckerMove.Analysis) > 0 {
-						err = d.saveCheckerAnalysisToPositionInTx(tx, posID, move.CheckerMove.Analysis,
-							&move.CheckerMove.Position, move.CheckerMove.ActivePlayer, &move.CheckerMove.PlayedMove)
-						if err != nil {
-							slog.Warn("failed to save analysis for canonical duplicate", "err", err)
-						}
-					}
-
-					// Save comment from XG file if present
-					if move.Comment != "" {
-						_, err = tx.Exec(`INSERT INTO comment (position_id, text) VALUES (?, ?)`, posID, move.Comment)
-						if err != nil {
-							slog.Warn("failed to save XG comment for canonical duplicate", "err", err)
-						}
-					}
-				} else if move.MoveType == "cube" && move.CubeMove != nil && move.CubeMove.Analysis != nil {
-					pos, err := d.createPositionFromXG(move.CubeMove.Position, &game, int32(match.Metadata.MatchLength), 0, move.CubeMove.ActivePlayer)
-					if err != nil {
-						continue
-					}
-					pos.PlayerOnRoll = convertXGPlayerToBlunderDB(move.CubeMove.ActivePlayer)
-					pos.DecisionType = CubeAction
-					pos.Dice = [2]int{0, 0}
-
-					posID, err := d.findOrCreatePositionForCanonicalDuplicate(tx, pos, cache)
-					if err != nil {
-						continue
-					}
-
-					err = d.saveCubeAnalysisToPositionInTx(tx, posID, move.CubeMove.Analysis, d.convertCubeAction(move.CubeMove.CubeAction))
-					if err != nil {
-						slog.Warn("failed to save cube analysis for canonical duplicate", "err", err)
-					}
-
-					// Save comment from XG file if present
-					if move.Comment != "" {
-						_, err = tx.Exec(`INSERT INTO comment (position_id, text) VALUES (?, ?)`, posID, move.Comment)
-						if err != nil {
-							slog.Warn("failed to save XG comment for canonical duplicate", "err", err)
-						}
-					}
-				}
-			}
-		}
-	} else {
-		// Normal import: create game/move records and import everything
-		if err := d.importXGGamesAndMoves(ctx, tx, matchID, match, rawCubeInfo, cache); err != nil {
-			return 0, err
-		}
+	// Preserve the legacy GUI/CLI contract: an exact same-format re-import is a
+	// duplicate error (the ingest layer instead reports it as a silent skip). A
+	// cross-format canonical duplicate is still enriched in place and returns
+	// the existing match id without error, as before.
+	if res.Skipped {
+		_ = tx.Rollback()
+		return 0, ErrDuplicateMatch
 	}
-
-	// Commit transaction
-	if err = tx.Commit(); err != nil {
-		return 0, fmt.Errorf("failed to commit transaction: %w", err)
+	if err := tx.Commit(); err != nil {
+		return 0, err
 	}
-
-	slog.Info("imported match", "matchID", matchID, "games", len(match.Games), "file", filePath)
-	return matchID, nil
+	slog.Info("imported match", "matchID", res.MatchID, "file", filePath)
+	return res.MatchID, nil
 }
 
 // RawCubeAction stores raw cube action data from XG file
@@ -234,496 +62,6 @@ type RawCubeAction struct {
 	CubeB    int32
 	Position [26]int8
 	Doubled  *xgparser.EngineStructDoubleAction // Full cube analysis data
-}
-
-// importXGGamesAndMoves imports all games, moves, and analysis from an XG match
-func (d *Database) importXGGamesAndMoves(ctx context.Context, tx *sql.Tx, matchID int64, match *xgparser.Match, rawCubeInfo map[string]*RawCubeAction, cache *importCache) error {
-	for gameIdx, game := range match.Games {
-		gameID, err := d.importGame(tx, matchID, &game)
-		if err != nil {
-			return fmt.Errorf("failed to import game %d: %w", game.GameNumber, err)
-		}
-
-		cubeIdx := 0
-		var lastCubeAnalysis *RawCubeAction
-		var pendingCubeComment string // Comment from a skipped "No Double" cube decision
-
-		for moveIdx, move := range game.Moves {
-			// Cancellation check at the top of every move iteration.
-			if err := ctx.Err(); err != nil {
-				return err
-			}
-
-			var rawCube *RawCubeAction
-
-			if move.MoveType == "cube" {
-				key := fmt.Sprintf("%d_%d", gameIdx+1, cubeIdx)
-				if rc, ok := rawCubeInfo[key]; ok {
-					rawCube = rc
-					lastCubeAnalysis = rc
-				}
-				cubeIdx++
-
-				// Check if this is a "No Double" that will be skipped by importMoveWithCacheAndRawCube.
-				// If it has a comment, carry it forward to the next checker move.
-				isNoDouble := false
-				if rawCube != nil {
-					isNoDouble = (rawCube.Double != 1)
-				} else if move.CubeMove != nil {
-					isNoDouble = (move.CubeMove.CubeAction == 0)
-				}
-				if isNoDouble && move.Comment != "" {
-					pendingCubeComment = move.Comment
-				}
-			} else if move.MoveType == "checker" {
-				rawCube = lastCubeAnalysis
-				lastCubeAnalysis = nil
-
-				// Carry forward comment from a skipped "No Double" cube decision
-				if pendingCubeComment != "" {
-					if move.Comment == "" {
-						move.Comment = pendingCubeComment
-					} else {
-						move.Comment = pendingCubeComment + "\n" + move.Comment
-					}
-					pendingCubeComment = ""
-				}
-			}
-
-			err := d.importMoveWithCacheAndRawCube(tx, gameID, int32(moveIdx), &move, &game, int32(match.Metadata.MatchLength), cache, rawCube)
-			if err != nil {
-				return fmt.Errorf("failed to import move %d in game %d: %w", moveIdx, game.GameNumber, err)
-			}
-		}
-	}
-	return nil
-}
-
-// importGame inserts a game and returns its ID
-func (d *Database) importGame(tx *sql.Tx, matchID int64, game *xgparser.Game) (int64, error) {
-	result, err := tx.Exec(`
-		INSERT INTO game (match_id, game_number, initial_score_1, initial_score_2,
-		                  winner, points_won, move_count)
-		VALUES (?, ?, ?, ?, ?, ?, ?)
-	`, matchID, game.GameNumber, game.InitialScore[0], game.InitialScore[1],
-		game.Winner, game.PointsWon, len(game.Moves))
-
-	if err != nil {
-		return 0, err
-	}
-
-	return result.LastInsertId()
-}
-
-// importMoveWithCacheAndRawCube imports a move with raw cube data for complete action info
-func (d *Database) importMoveWithCacheAndRawCube(tx *sql.Tx, gameID int64, moveNumber int32, move *xgparser.Move, game *xgparser.Game, matchLength int32, cache *importCache, rawCube *RawCubeAction) error {
-	var positionID int64
-	var player int32
-	var dice [2]int32
-	var checkerMoveStr string
-	var cubeActionStr string
-
-	if move.MoveType == "checker" && move.CheckerMove != nil {
-		// Create position from checker move
-		pos, err := d.createPositionFromXG(move.CheckerMove.Position, game, matchLength, moveNumber, move.CheckerMove.ActivePlayer)
-		if err != nil {
-			return fmt.Errorf("failed to create position: %w", err)
-		}
-
-		// Set position-specific attributes from move
-		// Convert XG player encoding (-1, 1) to blunderDB encoding (0, 1)
-		pos.PlayerOnRoll = convertXGPlayerToBlunderDB(move.CheckerMove.ActivePlayer)
-		pos.DecisionType = CheckerAction
-		pos.Dice = [2]int{int(move.CheckerMove.Dice[0]), int(move.CheckerMove.Dice[1])}
-
-		// Save position with cache
-		posID, err := d.savePositionInTxWithCache(tx, pos, cache)
-		if err != nil {
-			return fmt.Errorf("failed to save position: %w", err)
-		}
-		positionID = posID
-
-		player = move.CheckerMove.ActivePlayer
-		dice = move.CheckerMove.Dice
-
-		// Convert move notation with hit detection for consistency with analysis display
-		checkerMoveStr = d.convertXGMoveToStringWithHits(move.CheckerMove.PlayedMove, &move.CheckerMove.Position, move.CheckerMove.ActivePlayer)
-
-		// Save move
-		moveResult, err := tx.Exec(`
-			INSERT INTO move (game_id, move_number, move_type, position_id, player,
-			                  dice_1, dice_2, checker_move)
-			VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-		`, gameID, moveNumber, "checker", positionID, player, dice[0], dice[1], checkerMoveStr)
-		if err != nil {
-			return err
-		}
-
-		moveID, err := moveResult.LastInsertId()
-		if err != nil {
-			return fmt.Errorf("failed to get last insert ID: %w", err)
-		}
-
-		// Save analysis if available
-		if len(move.CheckerMove.Analysis) > 0 {
-			for _, analysis := range move.CheckerMove.Analysis {
-				err = d.saveMoveAnalysisInTx(tx, moveID, &analysis)
-				if err != nil {
-					slog.Warn("failed to save checker analysis", "err", err)
-				}
-			}
-
-			err = d.saveCheckerAnalysisToPositionInTx(tx, positionID, move.CheckerMove.Analysis, &move.CheckerMove.Position, move.CheckerMove.ActivePlayer, &move.CheckerMove.PlayedMove)
-			if err != nil {
-				slog.Warn("failed to save position analysis", "err", err)
-			}
-		}
-
-		// Save cube analysis for this checker position (from the preceding CubeEntry)
-		// This allows displaying cube info when pressing 'd' on a checker decision
-		if rawCube != nil && rawCube.Doubled != nil {
-			err = d.saveCubeAnalysisForCheckerPositionInTx(tx, positionID, rawCube)
-			if err != nil {
-				slog.Warn("failed to save cube analysis for checker position", "err", err)
-			}
-		}
-
-	} else if move.MoveType == "cube" && move.CubeMove != nil {
-		// Handle explicit cube decisions
-		// Only create positions for EXPLICIT cube actions (when a player actually doubles)
-		// Skip implicit "No Double" decisions
-
-		// Check if this is an explicit cube action
-		isExplicitCubeAction := false
-		if rawCube != nil {
-			// Explicit action: Double was offered (Double == 1)
-			isExplicitCubeAction = (rawCube.Double == 1)
-		} else {
-			// Fallback: check CubeAction field
-			isExplicitCubeAction = (move.CubeMove.CubeAction != 0) // 0 = No Double
-		}
-
-		if !isExplicitCubeAction {
-			// Create a No Double cube decision position when cube analysis is available.
-			// XG counts these in its "Cube Decisions" total, so we must too for accurate PR/MWC.
-			// Positions with no analysis are still skipped (nothing to compute).
-			hasRawAnalysis := rawCube != nil && rawCube.Doubled != nil
-			hasTextAnalysis := move.CubeMove.Analysis != nil
-			if !hasRawAnalysis && !hasTextAnalysis {
-				return nil // no analysis available — skip
-			}
-
-			// Only count as a cube decision if the player actually holds the cube.
-			// CubeB == 0 means centered (either player may double).
-			// CubeB > 0 means player 1 owns the cube (value = CubeB).
-			// CubeB < 0 means player 2 owns the cube (value = -CubeB).
-			// If the opponent owns the cube the active player has no cube decision.
-			if rawCube != nil {
-				canDouble := rawCube.CubeB == 0 ||
-					(rawCube.ActiveP > 0 && rawCube.CubeB > 0) ||
-					(rawCube.ActiveP < 0 && rawCube.CubeB < 0)
-				if !canDouble {
-					return nil
-				}
-			} else if move.CubeMove != nil {
-				// Fallback: use Position.Cube from the lightweight struct (same encoding)
-				cube := move.CubeMove.Position.Cube
-				ap := move.CubeMove.ActivePlayer
-				canDouble := cube == 0 || (ap > 0 && cube > 0) || (ap < 0 && cube < 0)
-				if !canDouble {
-					return nil
-				}
-			}
-
-			pos, err := d.createPositionFromXG(move.CubeMove.Position, game, matchLength, moveNumber, move.CubeMove.ActivePlayer)
-			if err != nil {
-				return nil // skip on position creation error
-			}
-			pos.PlayerOnRoll = convertXGPlayerToBlunderDB(move.CubeMove.ActivePlayer)
-			pos.DecisionType = CubeAction
-			pos.Dice = [2]int{0, 0}
-
-			posID, err := d.savePositionInTxWithCache(tx, pos, cache)
-			if err != nil {
-				return fmt.Errorf("failed to save no-double position: %w", err)
-			}
-
-			_, err = tx.Exec(`
-				INSERT INTO move (game_id, move_number, move_type, position_id, player,
-				                  dice_1, dice_2, cube_action)
-				VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-			`, gameID, moveNumber, "cube", posID, move.CubeMove.ActivePlayer, 0, 0, "No Double")
-			if err != nil {
-				return err
-			}
-
-			if hasRawAnalysis {
-				if err = d.saveRawCubeAnalysisForNoDoubleInTx(tx, posID, rawCube); err != nil {
-					slog.Warn("failed to save no-double raw analysis", "err", err)
-				}
-			} else {
-				if err = d.saveCubeAnalysisToPositionInTx(tx, posID, move.CubeMove.Analysis, "No Double"); err != nil {
-					slog.Warn("failed to save no-double text analysis", "err", err)
-				}
-			}
-			return nil
-		}
-
-		if rawCube != nil && rawCube.Double == 1 && rawCube.Take == 1 {
-			// DOUBLE/TAKE scenario: Create two positions
-
-			// Position 1: Doubling decision (before the double)
-			// The player on roll decides whether to double
-			pos1, err := d.createPositionFromXG(move.CubeMove.Position, game, matchLength, moveNumber, move.CubeMove.ActivePlayer)
-			if err != nil {
-				return fmt.Errorf("failed to create doubling position: %w", err)
-			}
-			pos1.PlayerOnRoll = convertXGPlayerToBlunderDB(move.CubeMove.ActivePlayer)
-			pos1.DecisionType = CubeAction
-			pos1.Dice = [2]int{0, 0}
-
-			// Save position 1 (doubling decision)
-			posID1, err := d.savePositionInTxWithCache(tx, pos1, cache)
-			if err != nil {
-				return fmt.Errorf("failed to save doubling position: %w", err)
-			}
-
-			player = move.CubeMove.ActivePlayer
-
-			// Save move 1: Double
-			moveResult1, err := tx.Exec(`
-				INSERT INTO move (game_id, move_number, move_type, position_id, player,
-				                  dice_1, dice_2, cube_action)
-				VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-			`, gameID, moveNumber, "cube", posID1, player, 0, 0, "Double")
-			if err != nil {
-				return err
-			}
-			moveID1, err := moveResult1.LastInsertId()
-			if err != nil {
-				return fmt.Errorf("failed to get last insert ID: %w", err)
-			}
-
-			// Save analysis to first position
-			if move.CubeMove.Analysis != nil {
-				err = d.saveCubeAnalysisInTx(tx, moveID1, move.CubeMove.Analysis)
-				if err != nil {
-					slog.Warn("failed to save cube analysis", "err", err)
-				}
-				err = d.saveCubeAnalysisToPositionInTx(tx, posID1, move.CubeMove.Analysis, "Double/Take")
-				if err != nil {
-					slog.Warn("failed to save position cube analysis", "err", err)
-				}
-			}
-
-			// Position 2: Take/Pass decision (after the double)
-			// The opponent decides whether to take or pass
-			// Note: We show the position BEFORE the take decision is executed,
-			// so the cube is still in the center at doubled value.
-			// The cube ownership will be reflected in the NEXT checker move position.
-			pos2 := *pos1                           // Clone the position
-			pos2.Cube.Value++                       // Double the cube (increment exponent: 0→1, 1→2, etc.)
-			opponentPlayer := 1 - pos1.PlayerOnRoll // Opponent player (blunderDB encoding)
-			pos2.PlayerOnRoll = opponentPlayer      // Opponent decides whether to take
-			pos2.Cube.Owner = -1                    // Cube still in center (decision not yet executed)
-
-			// Save position 2 (take decision)
-			posID2, err := d.savePositionInTxWithCache(tx, &pos2, cache)
-			if err != nil {
-				return fmt.Errorf("failed to save take position: %w", err)
-			}
-			positionID = posID2 // Use second position as the reference
-
-			// Convert opponent from blunderDB (0,1) back to XG encoding (1,-1) for move table
-			var opponentPlayerXG int32
-			if opponentPlayer == 0 {
-				opponentPlayerXG = 1
-			} else {
-				opponentPlayerXG = -1
-			}
-
-			// Save move 2: Take
-			_, err = tx.Exec(`
-				INSERT INTO move (game_id, move_number, move_type, position_id, player,
-				                  dice_1, dice_2, cube_action)
-				VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-			`, gameID, moveNumber, "cube", posID2, opponentPlayerXG, 0, 0, "Take")
-			if err != nil {
-				return err
-			}
-
-			// Save analysis to posID2 (take/pass decision) using the same cube analysis.
-			// The opponent decided to Take, so playedCubeAction = "Take".
-			if move.CubeMove.Analysis != nil {
-				err = d.saveCubeAnalysisToPositionInTx(tx, posID2, move.CubeMove.Analysis, "Take")
-				if err != nil {
-					slog.Warn("failed to save cube analysis for take position", "err", err)
-				}
-			}
-
-		} else {
-			// Single cube action (Double without Take, or other actions)
-			pos, err := d.createPositionFromXG(move.CubeMove.Position, game, matchLength, moveNumber, move.CubeMove.ActivePlayer)
-			if err != nil {
-				return fmt.Errorf("failed to create position: %w", err)
-			}
-
-			pos.PlayerOnRoll = convertXGPlayerToBlunderDB(move.CubeMove.ActivePlayer)
-			pos.DecisionType = CubeAction
-			pos.Dice = [2]int{0, 0}
-
-			// Save position
-			posID, err := d.savePositionInTxWithCache(tx, pos, cache)
-			if err != nil {
-				return fmt.Errorf("failed to save position: %w", err)
-			}
-			positionID = posID
-			player = move.CubeMove.ActivePlayer
-
-			// Determine cube action string
-			if rawCube != nil {
-				cubeActionStr = d.convertRawCubeAction(rawCube.Double, rawCube.Take)
-			} else {
-				cubeActionStr = d.convertCubeAction(move.CubeMove.CubeAction)
-			}
-
-			// Save move
-			moveResult, err := tx.Exec(`
-				INSERT INTO move (game_id, move_number, move_type, position_id, player,
-				                  dice_1, dice_2, cube_action)
-				VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-			`, gameID, moveNumber, "cube", positionID, player, 0, 0, cubeActionStr)
-			if err != nil {
-				return err
-			}
-
-			moveID, err := moveResult.LastInsertId()
-			if err != nil {
-				return fmt.Errorf("failed to get last insert ID: %w", err)
-			}
-
-			// Save cube analysis
-			if move.CubeMove.Analysis != nil {
-				err = d.saveCubeAnalysisInTx(tx, moveID, move.CubeMove.Analysis)
-				if err != nil {
-					slog.Warn("failed to save cube analysis", "err", err)
-				}
-
-				err = d.saveCubeAnalysisToPositionInTx(tx, positionID, move.CubeMove.Analysis, cubeActionStr)
-				if err != nil {
-					slog.Warn("failed to save position cube analysis", "err", err)
-				}
-			}
-
-			// For Double/Pass, also create the passer's take/pass decision position.
-			// In XG rawCube encoding, Take == 0 means the opponent passed the double.
-			// This mirrors what we do for Double/Take above so that the passer's
-			// decision is counted in per-player statistics.
-			if rawCube != nil && rawCube.Take == 0 {
-				pos2 := *pos
-				pos2.Cube.Value++                      // cube value doubles
-				opponentPlayer := 1 - pos.PlayerOnRoll // opponent (passer) in blunderDB encoding
-				pos2.PlayerOnRoll = opponentPlayer
-				pos2.Cube.Owner = -1 // cube still in centre during the take/pass decision
-
-				posID2, err2 := d.savePositionInTxWithCache(tx, &pos2, cache)
-				if err2 != nil {
-					slog.Warn("failed to save pass position", "err", err2)
-				} else {
-					var oppPlayerXG int32
-					if opponentPlayer == 0 {
-						oppPlayerXG = 1
-					} else {
-						oppPlayerXG = -1
-					}
-					_, err2 = tx.Exec(`
-						INSERT INTO move (game_id, move_number, move_type, position_id, player,
-						                  dice_1, dice_2, cube_action)
-						VALUES (?, ?, ?, ?, ?, ?, ?, ?)
-					`, gameID, moveNumber, "cube", posID2, oppPlayerXG, 0, 0, "Pass")
-					if err2 != nil {
-						slog.Warn("failed to insert pass move", "err", err2)
-					}
-					if move.CubeMove.Analysis != nil {
-						err2 = d.saveCubeAnalysisToPositionInTx(tx, posID2, move.CubeMove.Analysis, "Pass")
-						if err2 != nil {
-							slog.Warn("failed to save pass cube analysis", "err", err2)
-						}
-					}
-				}
-			}
-		}
-	}
-
-	// Save comment from XG file if present
-	if move.Comment != "" && positionID > 0 {
-		_, err := tx.Exec(`INSERT INTO comment (position_id, text) VALUES (?, ?)`, positionID, move.Comment)
-		if err != nil {
-			slog.Warn("failed to save XG comment", "err", err)
-		}
-	}
-
-	return nil
-}
-
-// saveRawCubeAnalysisForNoDoubleInTx saves cube analysis from binary (rawCube) data
-// to a standalone No Double cube decision position. Unlike
-// saveCubeAnalysisForCheckerPositionInTx (which merges onto a checker position),
-// this creates a "DoublingCube" analysis with playedCubeAction = "No Double" so
-// that populateAnalysisColumns computes the correct cube_error (missed-double error).
-func (d *Database) saveRawCubeAnalysisForNoDoubleInTx(tx *sql.Tx, positionID int64, rawCube *RawCubeAction) error {
-	if rawCube == nil || rawCube.Doubled == nil {
-		return nil
-	}
-	doubled := rawCube.Doubled
-
-	posAnalysis := PositionAnalysis{
-		PositionID:            int(positionID),
-		AnalysisType:          "DoublingCube",
-		AnalysisEngineVersion: "XG",
-		PlayedCubeActions:     []string{"No Double"},
-		CreationDate:          time.Now(),
-		LastModifiedDate:      time.Now(),
-	}
-
-	cubeAnalysis := buildDoublingCubeAnalysis(cubeAnalysisParams{
-		Depth:                     translateAnalysisDepth(int(doubled.Level)),
-		Engine:                    "XG",
-		PlayerWinChances:          (1.0 - float64(doubled.Eval[2])) * 100.0,
-		PlayerGammonChances:       float64(doubled.Eval[4]) * 100.0,
-		PlayerBackgammonChances:   float64(doubled.Eval[5]) * 100.0,
-		OpponentWinChances:        float64(doubled.Eval[2]) * 100.0,
-		OpponentGammonChances:     float64(doubled.Eval[1]) * 100.0,
-		OpponentBackgammonChances: float64(doubled.Eval[0]) * 100.0,
-		CubelessNoDoubleEquity:    float64(doubled.Eval[6]),
-		CubelessDoubleEquity:      float64(doubled.Eval[6]),
-		CubefulNoDoubleEquity:     float64(doubled.EquB),
-		CubefulDoubleTakeEquity:   float64(doubled.EquDouble),
-		CubefulDoublePassEquity:   float64(doubled.EquDrop),
-	})
-
-	posAnalysis.DoublingCubeAnalysis = &cubeAnalysis
-	return d.saveAnalysisInTx(tx, positionID, posAnalysis)
-}
-
-// convertRawCubeAction converts raw Double/Take values to action string
-func (d *Database) convertRawCubeAction(double, take int32) string {
-	// XG cube action encoding:
-	// Double=0, Take=-1: No Double
-	// Double=1, Take=1:  Double/Take (opponent took)
-	// Double=1, Take=0:  Double/Pass (opponent passed)
-	// Double=-2: Initial position (should be filtered before)
-
-	if double == 0 {
-		return "No Double"
-	} else if double == 1 {
-		if take == 1 {
-			return "Double/Take"
-		} else {
-			return "Double/Pass"
-		}
-	}
-	return fmt.Sprintf("Unknown(D=%d,T=%d)", double, take)
 }
 
 // createPositionFromXG converts xgparser.Position to blunderDB Position
@@ -903,38 +241,6 @@ func convertBlunderDBPlayerToXG(blunderDBPlayer int) int32 {
 	return -1 // Player 2
 }
 
-// saveMoveAnalysisInTx saves checker move analysis within a transaction
-func (d *Database) saveMoveAnalysisInTx(tx *sql.Tx, moveID int64, analysis *xgparser.CheckerAnalysis) error {
-	return insertMoveAnalysisRow(tx, moveAnalysisRow{
-		MoveID:                 moveID,
-		AnalysisType:           "checker",
-		Depth:                  translateAnalysisDepth(int(analysis.AnalysisDepth)),
-		Equity:                 int64(math.Round(float64(analysis.Equity) * 1000)),
-		WinRate:                int64(math.Round(float64(analysis.Player1WinRate) * 100.0 * 100)),
-		GammonRate:             int64(math.Round(float64(analysis.Player1GammonRate) * 100.0 * 100)),
-		BackgammonRate:         int64(math.Round(float64(analysis.Player1BgRate) * 100.0 * 100)),
-		OpponentWinRate:        int64(math.Round(float64(1.0-analysis.Player1WinRate) * 100.0 * 100)),
-		OpponentGammonRate:     int64(math.Round(float64(analysis.Player2GammonRate) * 100.0 * 100)),
-		OpponentBackgammonRate: int64(math.Round(float64(analysis.Player2BgRate) * 100.0 * 100)),
-	})
-}
-
-// saveCubeAnalysisInTx saves cube analysis within a transaction
-func (d *Database) saveCubeAnalysisInTx(tx *sql.Tx, moveID int64, analysis *xgparser.CubeAnalysis) error {
-	return insertMoveAnalysisRow(tx, moveAnalysisRow{
-		MoveID:                 moveID,
-		AnalysisType:           "cube",
-		Depth:                  translateAnalysisDepth(int(analysis.AnalysisDepth)),
-		Equity:                 int64(math.Round(float64(analysis.CubefulNoDouble) * 1000)),
-		WinRate:                int64(math.Round(float64(analysis.Player1WinRate) * 100.0 * 100)),
-		GammonRate:             int64(math.Round(float64(analysis.Player1GammonRate) * 100.0 * 100)),
-		BackgammonRate:         int64(math.Round(float64(analysis.Player1BgRate) * 100.0 * 100)),
-		OpponentWinRate:        int64(math.Round(float64(1.0-analysis.Player1WinRate) * 100.0 * 100)),
-		OpponentGammonRate:     int64(math.Round(float64(analysis.Player2GammonRate) * 100.0 * 100)),
-		OpponentBackgammonRate: int64(math.Round(float64(analysis.Player2BgRate) * 100.0 * 100)),
-	})
-}
-
 // saveCheckerAnalysisToPositionInTx converts XG checker analysis to PositionAnalysis and saves it
 // playedMove is optional - if provided, it will be used as the source of truth for the first analysis
 // (workaround for xgparser bug where analysis.Move may be incomplete for multi-submove bear-offs)
@@ -1082,55 +388,6 @@ func (d *Database) saveCubeAnalysisToPositionInTx(tx *sql.Tx, positionID int64, 
 	posAnalysis.DoublingCubeAnalysis = &cubeAnalysis
 
 	// Save to analysis table
-	return d.saveAnalysisInTx(tx, positionID, posAnalysis)
-}
-
-// saveCubeAnalysisForCheckerPositionInTx saves cube analysis from a RawCubeAction to a checker position
-// This is used to attach the cube decision analysis to checker moves (from the preceding CubeEntry)
-// It merges the cube info with existing checker analysis if present
-func (d *Database) saveCubeAnalysisForCheckerPositionInTx(tx *sql.Tx, positionID int64, rawCube *RawCubeAction) error {
-	if rawCube == nil || rawCube.Doubled == nil {
-		return nil
-	}
-
-	doubled := rawCube.Doubled
-
-	// Build a PositionAnalysis with just the cube analysis and let saveAnalysisInTx handle merging
-	posAnalysis := PositionAnalysis{
-		PositionID:            int(positionID),
-		AnalysisType:          "CheckerMove",
-		AnalysisEngineVersion: "XG",
-		CreationDate:          time.Now(),
-		LastModifiedDate:      time.Now(),
-	}
-
-	// Extract data from EngineStructDoubleAction
-	// XG Eval array mapping (from xgparser convertCubeEntry):
-	// Eval[0] = opponent's backgammon rate
-	// Eval[1] = opponent's gammon rate
-	// Eval[2] = opponent's win rate (so player on roll's win rate = 1 - Eval[2])
-	// Eval[4] = player on roll's gammon rate
-	// Eval[5] = player on roll's backgammon rate
-	// Eval[6] = cubeless equity
-	cubeAnalysis := buildDoublingCubeAnalysis(cubeAnalysisParams{
-		Depth:                     translateAnalysisDepth(int(doubled.Level)),
-		Engine:                    "XG",
-		PlayerWinChances:          (1.0 - float64(doubled.Eval[2])) * 100.0,
-		PlayerGammonChances:       float64(doubled.Eval[4]) * 100.0,
-		PlayerBackgammonChances:   float64(doubled.Eval[5]) * 100.0,
-		OpponentWinChances:        float64(doubled.Eval[2]) * 100.0,
-		OpponentGammonChances:     float64(doubled.Eval[1]) * 100.0,
-		OpponentBackgammonChances: float64(doubled.Eval[0]) * 100.0,
-		CubelessNoDoubleEquity:    float64(doubled.Eval[6]),
-		CubelessDoubleEquity:      float64(doubled.Eval[6]),
-		CubefulNoDoubleEquity:     float64(doubled.EquB),
-		CubefulDoubleTakeEquity:   float64(doubled.EquDouble),
-		CubefulDoublePassEquity:   float64(doubled.EquDrop),
-	})
-
-	posAnalysis.DoublingCubeAnalysis = &cubeAnalysis
-
-	// Save to analysis table (saveAnalysisInTx handles merge with existing)
 	return d.saveAnalysisInTx(tx, positionID, posAnalysis)
 }
